backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
python-jose[cryptography]==3.4.0
python-multipart==0.0.20
gradio-client==1.10.1
pydantic==2.11.3
pydantic-settings==2.9.1
python-dotenv==1.1.0
aiofiles==24.1.0
